# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Student Portal API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type AnnouncementPriority =
  (typeof AnnouncementPriority)[keyof typeof AnnouncementPriority];

export const AnnouncementPriority = {
  low: "low",
  normal: "normal",
  high: "high",
  urgent: "urgent",
} as const;

export interface Announcement {
  id: number;
  title: string;
  content: string;
  category: string;
  priority: AnnouncementPriority;
  authorName: string;
  isActive: boolean;
  createdAt: string;
  updatedAt: string;
}

export type CreateAnnouncementBodyPriority =
  (typeof CreateAnnouncementBodyPriority)[keyof typeof CreateAnnouncementBodyPriority];

export const CreateAnnouncementBodyPriority = {
  low: "low",
  normal: "normal",
  high: "high",
  urgent: "urgent",
} as const;

export interface CreateAnnouncementBody {
  title: string;
  content: string;
  category: string;
  priority: CreateAnnouncementBodyPriority;
  authorName: string;
}

export type AnnouncementsSummaryByCategoryItem = {
  category: string;
  count: number;
};

export interface AnnouncementsSummary {
  total: number;
  urgent: number;
  byCategory: AnnouncementsSummaryByCategoryItem[];
}

export type DocumentStatus =
  (typeof DocumentStatus)[keyof typeof DocumentStatus];

export const DocumentStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;

export interface Document {
  id: number;
  studentId: number;
  studentName: string;
  fileName: string;
  fileType: string;
  fileSize: number;
  documentType: string;
  description?: string;
  status: DocumentStatus;
  createdAt: string;
}

export interface CreateDocumentBody {
  studentId: number;
  studentName: string;
  fileName: string;
  fileType: string;
  fileSize: number;
  documentType: string;
  description?: string;
}

export type StudentStatus = (typeof StudentStatus)[keyof typeof StudentStatus];

export const StudentStatus = {
  active: "active",
  inactive: "inactive",
  graduated: "graduated",
} as const;

export interface Student {
  id: number;
  matricula: string;
  firstName: string;
  lastName: string;
  email: string;
  program: string;
  semester: number;
  campus: string;
  status: StudentStatus;
  createdAt: string;
}

export interface CreateStudentBody {
  matricula: string;
  firstName: string;
  lastName: string;
  email: string;
  program: string;
  semester: number;
  campus: string;
}

export interface Course {
  id: number;
  code: string;
  name: string;
  professor: string;
  credits: number;
  schedule: string;
  room: string;
  semester: string;
  createdAt: string;
}

export interface CreateCourseBody {
  code: string;
  name: string;
  professor: string;
  credits: number;
  schedule: string;
  room: string;
  semester: string;
}

export interface DashboardSummary {
  totalStudents: number;
  activeStudents: number;
  totalAnnouncements: number;
  urgentAnnouncements: number;
  totalDocuments: number;
  pendingDocuments: number;
  totalCourses: number;
  recentAnnouncements: Announcement[];
  recentDocuments: Document[];
}

export type ListAnnouncementsParams = {
  category?: string;
  limit?: number;
};

export type ListDocumentsParams = {
  studentId?: number;
  type?: string;
};