# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Student Portal API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: announcements
    description: Announcement operations
  - name: documents
    description: Document upload operations
  - name: students
    description: Student profile operations
  - name: courses
    description: Course operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /announcements:
    get:
      operationId: listAnnouncements
      tags: [announcements]
      summary: List all announcements
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of announcements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Announcement"
    post:
      operationId: createAnnouncement
      tags: [announcements]
      summary: Create announcement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAnnouncementBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"

  /announcements/{id}:
    get:
      operationId: getAnnouncement
      tags: [announcements]
      summary: Get announcement by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Announcement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
        "404":
          description: Not found

  /announcements/summary:
    get:
      operationId: getAnnouncementsSummary
      tags: [announcements]
      summary: Get announcements summary stats
      responses:
        "200":
          description: Summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnnouncementsSummary"

  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List all uploaded documents
      parameters:
        - name: studentId
          in: query
          schema:
            type: integer
        - name: type
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
    post:
      operationId: createDocument
      tags: [documents]
      summary: Upload document metadata
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDocumentBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"

  /documents/{id}:
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /students:
    get:
      operationId: listStudents
      tags: [students]
      summary: List all students
      responses:
        "200":
          description: List of students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Student"
    post:
      operationId: createStudent
      tags: [students]
      summary: Create student profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStudentBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"

  /students/{id}:
    get:
      operationId: getStudent
      tags: [students]
      summary: Get student by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "404":
          description: Not found

  /courses:
    get:
      operationId: listCourses
      tags: [courses]
      summary: List all courses
      responses:
        "200":
          description: List of courses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Course"
    post:
      operationId: createCourse
      tags: [courses]
      summary: Create a course
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCourseBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"

  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [students]
      summary: Get dashboard summary (counts, recent activity)
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Announcement:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        category:
          type: string
        priority:
          type: string
          enum: [low, normal, high, urgent]
        authorName:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - content
        - category
        - priority
        - authorName
        - isActive
        - createdAt
        - updatedAt

    CreateAnnouncementBody:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        category:
          type: string
        priority:
          type: string
          enum: [low, normal, high, urgent]
        authorName:
          type: string
      required:
        - title
        - content
        - category
        - priority
        - authorName

    AnnouncementsSummary:
      type: object
      properties:
        total:
          type: integer
        urgent:
          type: integer
        byCategory:
          type: array
          items:
            type: object
            properties:
              category:
                type: string
              count:
                type: integer
            required:
              - category
              - count
      required:
        - total
        - urgent
        - byCategory

    Document:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: integer
        studentName:
          type: string
        fileName:
          type: string
        fileType:
          type: string
        fileSize:
          type: integer
        documentType:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - studentId
        - studentName
        - fileName
        - fileType
        - fileSize
        - documentType
        - status
        - createdAt

    CreateDocumentBody:
      type: object
      properties:
        studentId:
          type: integer
        studentName:
          type: string
        fileName:
          type: string
        fileType:
          type: string
        fileSize:
          type: integer
        documentType:
          type: string
        description:
          type: string
      required:
        - studentId
        - studentName
        - fileName
        - fileType
        - fileSize
        - documentType

    Student:
      type: object
      properties:
        id:
          type: integer
        matricula:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        program:
          type: string
        semester:
          type: integer
        campus:
          type: string
        status:
          type: string
          enum: [active, inactive, graduated]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - matricula
        - firstName
        - lastName
        - email
        - program
        - semester
        - campus
        - status
        - createdAt

    CreateStudentBody:
      type: object
      properties:
        matricula:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        program:
          type: string
        semester:
          type: integer
        campus:
          type: string
      required:
        - matricula
        - firstName
        - lastName
        - email
        - program
        - semester
        - campus

    Course:
      type: object
      properties:
        id:
          type: integer
        code:
          type: string
        name:
          type: string
        professor:
          type: string
        credits:
          type: integer
        schedule:
          type: string
        room:
          type: string
        semester:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - code
        - name
        - professor
        - credits
        - schedule
        - room
        - semester
        - createdAt

    CreateCourseBody:
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        professor:
          type: string
        credits:
          type: integer
        schedule:
          type: string
        room:
          type: string
        semester:
          type: string
      required:
        - code
        - name
        - professor
        - credits
        - schedule
        - room
        - semester

    DashboardSummary:
      type: object
      properties:
        totalStudents:
          type: integer
        activeStudents:
          type: integer
        totalAnnouncements:
          type: integer
        urgentAnnouncements:
          type: integer
        totalDocuments:
          type: integer
        pendingDocuments:
          type: integer
        totalCourses:
          type: integer
        recentAnnouncements:
          type: array
          items:
            $ref: "#/components/schemas/Announcement"
        recentDocuments:
          type: array
          items:
            $ref: "#/components/schemas/Document"
      required:
        - totalStudents
        - activeStudents
        - totalAnnouncements
        - urgentAnnouncements
        - totalDocuments
        - pendingDocuments
        - totalCourses
        - recentAnnouncements
        - recentDocuments