# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Student Portal API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all announcements
 */
export const listAnnouncementsQueryLimitDefault = 20;

export const ListAnnouncementsQueryParams = zod.object({
  category: zod.coerce.string().optional(),
  limit: zod.coerce.number().default(listAnnouncementsQueryLimitDefault),
});

export const ListAnnouncementsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  content: zod.string(),
  category: zod.string(),
  priority: zod.enum(["low", "normal", "high", "urgent"]),
  authorName: zod.string(),
  isActive: zod.boolean(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});
export const ListAnnouncementsResponse = zod.array(
  ListAnnouncementsResponseItem,
);

/**
 * @summary Create announcement
 */
export const CreateAnnouncementBody = zod.object({
  title: zod.string(),
  content: zod.string(),
  category: zod.string(),
  priority: zod.enum(["low", "normal", "high", "urgent"]),
  authorName: zod.string(),
});

/**
 * @summary Get announcement by ID
 */
export const GetAnnouncementParams = zod.object({
  id: zod.coerce.number(),
});

export const GetAnnouncementResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  content: zod.string(),
  category: zod.string(),
  priority: zod.enum(["low", "normal", "high", "urgent"]),
  authorName: zod.string(),
  isActive: zod.boolean(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Get announcements summary stats
 */
export const GetAnnouncementsSummaryResponse = zod.object({
  total: zod.number(),
  urgent: zod.number(),
  byCategory: zod.array(
    zod.object({
      category: zod.string(),
      count: zod.number(),
    }),
  ),
});

/**
 * @summary List all uploaded documents
 */
export const ListDocumentsQueryParams = zod.object({
  studentId: zod.coerce.number().optional(),
  type: zod.coerce.string().optional(),
});

export const ListDocumentsResponseItem = zod.object({
  id: zod.number(),
  studentId: zod.number(),
  studentName: zod.string(),
  fileName: zod.string(),
  fileType: zod.string(),
  fileSize: zod.number(),
  documentType: zod.string(),
  description: zod.string().optional(),
  status: zod.enum(["pending", "approved", "rejected"]),
  createdAt: zod.coerce.date(),
});
export const ListDocumentsResponse = zod.array(ListDocumentsResponseItem);

/**
 * @summary Upload document metadata
 */
export const CreateDocumentBody = zod.object({
  studentId: zod.number(),
  studentName: zod.string(),
  fileName: zod.string(),
  fileType: zod.string(),
  fileSize: zod.number(),
  documentType: zod.string(),
  description: zod.string().optional(),
});

/**
 * @summary Delete a document
 */
export const DeleteDocumentParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List all students
 */
export const ListStudentsResponseItem = zod.object({
  id: zod.number(),
  matricula: zod.string(),
  firstName: zod.string(),
  lastName: zod.string(),
  email: zod.string(),
  program: zod.string(),
  semester: zod.number(),
  campus: zod.string(),
  status: zod.enum(["active", "inactive", "graduated"]),
  createdAt: zod.coerce.date(),
});
export const ListStudentsResponse = zod.array(ListStudentsResponseItem);

/**
 * @summary Create student profile
 */
export const CreateStudentBody = zod.object({
  matricula: zod.string(),
  firstName: zod.string(),
  lastName: zod.string(),
  email: zod.string(),
  program: zod.string(),
  semester: zod.number(),
  campus: zod.string(),
});

/**
 * @summary Get student by ID
 */
export const GetStudentParams = zod.object({
  id: zod.coerce.number(),
});

export const GetStudentResponse = zod.object({
  id: zod.number(),
  matricula: zod.string(),
  firstName: zod.string(),
  lastName: zod.string(),
  email: zod.string(),
  program: zod.string(),
  semester: zod.number(),
  campus: zod.string(),
  status: zod.enum(["active", "inactive", "graduated"]),
  createdAt: zod.coerce.date(),
});

/**
 * @summary List all courses
 */
export const ListCoursesResponseItem = zod.object({
  id: zod.number(),
  code: zod.string(),
  name: zod.string(),
  professor: zod.string(),
  credits: zod.number(),
  schedule: zod.string(),
  room: zod.string(),
  semester: zod.string(),
  createdAt: zod.coerce.date(),
});
export const ListCoursesResponse = zod.array(ListCoursesResponseItem);

/**
 * @summary Create a course
 */
export const CreateCourseBody = zod.object({
  code: zod.string(),
  name: zod.string(),
  professor: zod.string(),
  credits: zod.number(),
  schedule: zod.string(),
  room: zod.string(),
  semester: zod.string(),
});

/**
 * @summary Get dashboard summary (counts, recent activity)
 */
export const GetDashboardSummaryResponse = zod.object({
  totalStudents: zod.number(),
  activeStudents: zod.number(),
  totalAnnouncements: zod.number(),
  urgentAnnouncements: zod.number(),
  totalDocuments: zod.number(),
  pendingDocuments: zod.number(),
  totalCourses: zod.number(),
  recentAnnouncements: zod.array(
    zod.object({
      id: zod.number(),
      title: zod.string(),
      content: zod.string(),
      category: zod.string(),
      priority: zod.enum(["low", "normal", "high", "urgent"]),
      authorName: zod.string(),
      isActive: zod.boolean(),
      createdAt: zod.coerce.date(),
      updatedAt: zod.coerce.date(),
    }),
  ),
  recentDocuments: zod.array(
    zod.object({
      id: zod.number(),
      studentId: zod.number(),
      studentName: zod.string(),
      fileName: zod.string(),
      fileType: zod.string(),
      fileSize: zod.number(),
      documentType: zod.string(),
      description: zod.string().optional(),
      status: zod.enum(["pending", "approved", "rejected"]),
      createdAt: zod.coerce.date(),
    }),
  ),
});