# lib/api-zod/src/generated/types/announcement.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Student Portal API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AnnouncementPriority } from "./announcementPriority";

export interface Announcement {
  id: number;
  title: string;
  content: string;
  category: string;
  priority: AnnouncementPriority;
  authorName: string;
  isActive: boolean;
  createdAt: Date;
  updatedAt: Date;
}