# lib/api-zod/src/generated/types/announcementsSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Student Portal API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AnnouncementsSummaryByCategoryItem } from "./announcementsSummaryByCategoryItem";

export interface AnnouncementsSummary {
  total: number;
  urgent: number;
  byCategory: AnnouncementsSummaryByCategoryItem[];
}