# lib/api-zod/src/generated/types/createAnnouncementBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Student Portal API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateAnnouncementBodyPriority } from "./createAnnouncementBodyPriority";

export interface CreateAnnouncementBody {
  title: string;
  content: string;
  category: string;
  priority: CreateAnnouncementBodyPriority;
  authorName: string;
}