# lib/api-zod/src/generated/types/createStudentBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Student Portal API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateStudentBody {
  matricula: string;
  firstName: string;
  lastName: string;
  email: string;
  program: string;
  semester: number;
  campus: string;
}