# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Student Portal API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Announcement } from "./announcement";
import type { Document } from "./document";

export interface DashboardSummary {
  totalStudents: number;
  activeStudents: number;
  totalAnnouncements: number;
  urgentAnnouncements: number;
  totalDocuments: number;
  pendingDocuments: number;
  totalCourses: number;
  recentAnnouncements: Announcement[];
  recentDocuments: Document[];
}