# lib/api-zod/src/generated/types/document.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Student Portal API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DocumentStatus } from "./documentStatus";

export interface Document {
  id: number;
  studentId: number;
  studentName: string;
  fileName: string;
  fileType: string;
  fileSize: number;
  documentType: string;
  description?: string;
  status: DocumentStatus;
  createdAt: Date;
}