# lib/api-zod/src/generated/types/student.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Student Portal API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StudentStatus } from "./studentStatus";

export interface Student {
  id: number;
  matricula: string;
  firstName: string;
  lastName: string;
  email: string;
  program: string;
  semester: number;
  campus: string;
  status: StudentStatus;
  createdAt: Date;
}